This fixes all of the "High", "Error", and "Warnings" issues detected by CodeQL. However, there were some that I classified as "Won't Fix" and "False Positive".  But I'm not sure if CodeQL will remember those markings.

+++ src/main/java/org/apache/datasketches/cpc/PreambleUtil.java
@@ -790,7 +790,7 @@ final class PreambleUtil {
     final long memCap = mem.getCapacity();
     final long expectedCap = offsetBytes + (4L * lengthInts);
     checkCapacity(memCap, expectedCap);
-    for (int i = 0; i < lengthInts; i++) {
+    for (long i = 0; i < lengthInts; i++) {
       sb.append(String.format(fmt, i, mem.getInt(offsetBytes + (4L * i)))).append(LS);
     }
   }
